+++ lib/onlyoffice_webdriver_wrapper/webdriver.rb
@@ -168,7 +168,8 @@ module OnlyofficeWebdriverWrapper
         set_style_attribute("#{xpath_name}/button", 'display', 'inline-block')
         set_style_attribute(xpath_name, 'display', 'block')
       else
-        @driver.action.move_to(element, horizontal_shift, vertical_shift).click.perform
+        shift_to_zero = move_to_shift_to_top_left(xpath_name)
+        @driver.action.move_to(element, horizontal_shift - shift_to_zero.x, vertical_shift - shift_to_zero.y).click.perform
       end
     end
 
@@ -182,7 +183,10 @@ module OnlyofficeWebdriverWrapper
     def action_on_locator_coordinates(xpath_name, right_by, down_by, action = :click, times = 1)
       wait_until_element_visible(xpath_name)
       element = @driver.find_element(:xpath, xpath_name)
-      (0...times).inject(@driver.action.move_to(element, right_by.to_i, down_by.to_i)) { |acc, _elem| acc.send(action) }.perform
+      shift_to_zero = move_to_shift_to_top_left(xpath_name)
+      (0...times).inject(@driver.action.move_to(element,
+                                                right_by.to_i - shift_to_zero.x,
+                                                down_by.to_i - shift_to_zero.y)) { |acc, _elem| acc.send(action) }.perform
     end
 
     # Check if element present on page